requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-activity-stream==0.8.0
django-cors-headers==3.2.0
django-extensions==2.2.9
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2==2.8.4
pydotplus==2.0.2
pyparsing==2.4.7
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
